Started work on project page
Separated some of the styles on CSS just for clarity. Idea for projects. When you hover over them, they expand  little bit and show the name and type of the project. Once user clicks on them, it will bring up a whole new page with more details of the project and a download link (if applicable).