New pages for projects
Did a bit of reorganizing CSS to have a base CSS file and then the others to include what is on their specific pages on a different file. Also added new webpages for each project as well as a basic webpage  layout file that can be used when creating new pages later. Learned new navigation technique using <base> tag in the head to dictate the starting route for files since running the webpage locally takes my drive as a root folder instead of what the folder with the files being that root.