Quality of management update
Added 3 different HTML files that contain the HTML needed for making the header and foot of each page. Use Javascript to get the header and footer elements to apply the HTML in those elements. This allows me to only need to update one of those 3 html files to make changes across multiple pages instead of having to go to each one individually.